Refactored everything from the App component into multiple components
