+++ package.json
@@ -13,6 +13,7 @@
     "@monaco-editor/react": "^4.6.0",
     "@types/jscodeshift": "0.7.2",
     "@types/prettier": "^2.7.3",
+    "@types/react-resizable": "3.0.2",
     "axios": "^1.6.5",
     "bulmaswatch": "^0.8.1",
     "esbuild-wasm": "0.8.27",
@@ -21,7 +22,8 @@
     "monaco-jsx-highlighter": "0.0.15",
     "prettier": "^2.8.8",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-resizable": "3.0.4"
   },
   "devDependencies": {
     "@types/react": "^18.2.43",
