+++ src/App.tsx
@@ -1,84 +1,13 @@
-import * as esbuild from "esbuild-wasm";
-import { useState, useEffect, useRef } from "react";
 import "bulmaswatch/superhero/bulmaswatch.min.css";
 
-import { unpkgPathPlugin } from "./plugins/unpkg-path-plugin";
-import { fetchPlugin } from "./plugins/fetch-plugin";
-
-import CodeEditor from "./components/CodeEditor";
+import CodeCell from "./components/CodeCell";
 
 const App = () => {
-  const [input, setInput] = useState<string>("");
-
-  const ref = useRef<any>();
-  const iframe = useRef<any>();
-
-  const startService = async () => {
-    ref.current = await esbuild.startService({
-      worker: true,
-      wasmURL: "https://unpkg.com/esbuild-wasm@0.8.27/esbuild.wasm",
-    });
-  };
-
-  useEffect(() => {
-    startService();
-  }, []);
-
-  const onChange = (value: string) => setInput(value);
-
-  const onClick = async () => {
-    if (!ref.current) return;
-
-    iframe.current.srcdoc = html;
-
-    const result = await ref.current.build({
-      entryPoints: ["index.js"],
-      bundle: true,
-      write: false,
-      plugins: [unpkgPathPlugin(), fetchPlugin(input)],
-      define: {
-        "process.env.NODE_ENV": "'production'",
-        global: "window",
-      },
-    });
-
-    // setCode(result.outputFiles[0].text);
-    iframe.current?.contentWindow.postMessage(result.outputFiles[0].text, "*");
-  };
-
-  const html = `
-    <html>
-      <head>
-      </head>
-      <body>
-        <div id="root"></div>
-        <script>
-          window.addEventListener('message', (event) => {
-            try {
-              eval(event.data)
-            } catch (err) {
-              const root = document.getElementById("root")
-              root.innerHTML = "<div style='color: red;'><h4>Runtime Error:</h4>" + err + "</div>"
-              console.error(err)
-            }
-          }, false)
-        </script>
-      </body>
-    </html>
-  `;
-
   return (
     <div>
-      <CodeEditor initialValue={input} onChange={onChange} />
-      <div>
-        <button onClick={onClick}>Submit</button>
-      </div>
-      <iframe
-        title="preview"
-        ref={iframe}
-        srcDoc={html}
-        sandbox="allow-scripts"
-      />
+      <CodeCell />
+      <CodeCell />
+      <CodeCell />
     </div>
   );
 };
