+++ src/bundler/index.ts
@@ -1,0 +1,31 @@
+import * as esbuild from "esbuild-wasm";
+
+import { unpkgPathPlugin } from "../plugins/unpkg-path-plugin";
+import { fetchPlugin } from "../plugins/fetch-plugin";
+
+let service: esbuild.Service;
+
+export default async (rawCode: string) => {
+  if (!service) {
+    service = await esbuild.startService({
+      worker: true,
+      wasmURL: "https://unpkg.com/esbuild-wasm@0.8.27/esbuild.wasm",
+    });
+  }
+
+  const result = await service.build({
+    entryPoints: ["index.js"],
+    bundle: true,
+    write: false,
+    plugins: [
+      unpkgPathPlugin() as esbuild.Plugin,
+      fetchPlugin(rawCode) as esbuild.Plugin,
+    ],
+    define: {
+      "process.env.NODE_ENV": "'production'",
+      global: "window",
+    },
+  });
+
+  return result.outputFiles[0].text;
+};
