+++ src/components/CodeCell.tsx
@@ -1,0 +1,30 @@
+import { useState } from "react";
+
+import bundle from "../bundler";
+
+import CodeEditor from "../components/CodeEditor";
+import Preview from "../components/Preview";
+
+const CodeCell: React.FC = () => {
+  const [input, setInput] = useState<string>("");
+  const [code, setCode] = useState<string>("");
+
+  const onChange = (value: string) => setInput(value);
+
+  const onClick = async () => {
+    const output = await bundle(input);
+    setCode(output);
+  };
+
+  return (
+    <div>
+      <CodeEditor initialValue={input} onChange={onChange} />
+      <div>
+        <button onClick={onClick}>Submit</button>
+      </div>
+      <Preview code={code} />
+    </div>
+  );
+};
+
+export default CodeCell;
