+++ src/components/CodeEditor.tsx
@@ -3,7 +3,7 @@ import MonacoEditor from "@monaco-editor/react";
 import prettier from "prettier";
 import parser from "prettier/parser-babel";
 
-import "./code-editor.css";
+import "./CodeEditor.css";
 
 interface CodeEditorProps {
   initialValue: string;
